requirements.txt
numpy==1.26.3
langchain==0.3.1
langchain-openai==0.2.0
langchain-community==0.3.1
python-dotenv==1.0.1
streamlit==1.39.0
supabase==2.17.0
pypdf==5.0.1
tenacity==8.3.0
